requirements.txt
azure-functions
psycopg2
requests
azure-storage-blob
sklearn
pandas